rename: kakao 관련 디렉토리 변경
